% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_magicalized.R
\name{is_magicalized}
\alias{is_magicalized}
\title{Check whether for() is magicalized or not}
\usage{
is_magicalized(envir = parent.frame())
}
\arguments{
\item{envir}{environment}
}
\value{
TRUE if for() is magicalized.
}
\description{
Check whether for() is magicalized or not
}

