% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_for.R
\name{magic_for}
\alias{magic_for}
\title{Magicalize for()}
\usage{
magic_for(func = put, progress = FALSE, test = FALSE,
  max_object_size = 1 * MB)
}
\arguments{
\item{func}{a name.}

\item{progress}{logical.}

\item{max_object_size}{a number. default 1 MB.}
}
\description{
Magicalize for()
}

