% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_result.R
\name{magic_result_as_vector}
\alias{magic_result_as_vector}
\title{Get values stored by magicalized for loops as a vector}
\usage{
magic_result_as_vector()
}
\description{
Get values stored by magicalized for loops as a vector
}

