% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_for.R
\name{magic_for}
\alias{magic_for}
\title{Magicalize for()}
\usage{
magic_for(func = put, progress = FALSE, test = FALSE, silent = FALSE,
  temp = FALSE, max_object_size = 1 * MB)
}
\arguments{
\item{func}{function name. The target function for magicalization.}

\item{progress}{logical. If TRUE, show progress bar.}

\item{test}{logical or a number. If TRUE or a number, limit iteration times to it.}

\item{silent}{logical. If TRUE, do not execute func. Usually, func is print something.}

\item{temp}{logical. If TRUE, once run for(), free magicalization.}

\item{max_object_size}{a number. Prevent to store large iterator. Default to 1 MB.}
}
\description{
Magicalize for()
}

